+++ Libraries/LibWeb/Painting/ViewportPaintable.h
@@ -22,6 +22,8 @@ public:
     virtual void reset_for_relayout() override;
 
     void paint_all_phases(DisplayListRecordingContext&);
+    void initialize_async_scrolling_metadata_recording(DisplayListRecordingContext&);
+    void finalize_async_scrolling_metadata_recording(DisplayListRecordingContext&, HTML::Navigable&, Gfx::IntRect viewport_rect);
     void build_stacking_context_tree_if_needed();
 
     void assign_scroll_frames();
